Shared Driver Package Rename
Renamed the package for Shared driver.
Moved the package for MathUtils.
Other small fixes
