+++ src/test/java/pageObject/YoutubeVideoPage.java
@@ -1,13 +1,13 @@
 package pageObject;
 
+import driver.SharedDriver;
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import pageObject.Utils.MathUtils;
-import shareddriver.SharedDriver;
+import utils.MathUtils;
 
 public class YoutubeVideoPage extends AbstractPage {
